Move unused and unknown classes to test directory
The purpose of these classes is unknown. Since they are
not used and clearly marked as a temporary
implementation move them to the test directory (just
in case we need them in the future).
