+++ src/sardana/taurus/core/tango/sardana/sardana.py
@@ -26,24 +26,12 @@
 """The sardana submodule. It contains specific part of sardana"""
 
 __all__ = ["BaseSardanaElement", "BaseSardanaElementContainer",
-           "Pool", "MacroServer", "Door", "Sardana", "SardanaManager",
-           "PoolElementType", "ControllerClassInfo", "ControllerInfo",
-           "ChannelView", "PlotType", "Normalization", "AcqTriggerType",
-           "AcqMode"]
-
-#-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-
-# T E M P O R A R Y   I M P L E M E N T A T I O N
-#
-# THIS IS USED FOR TEST PURPOSES ONLY. DO NOT USE IT OUTSIDE SARDANA TESTS
-#-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-
+           "PoolElementType", "ChannelView", "PlotType", "Normalization",
+           "AcqTriggerType", "AcqMode"]
 
 __docformat__ = 'restructuredtext'
 
-import PyTango
-
 from taurus.core.util.enumeration import Enumeration
-from taurus.core.util.log import Logger
-from taurus.core.util.singleton import Singleton
 from taurus.core.util.containers import CaselessDict
 from taurus.core.util.codecs import CodecFactory
 
@@ -230,516 +218,3 @@ class BaseSardanaElementContainer:
 
     def getTypes(self):
         return self._type_elems_dict
-
-#-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-
-# T E M P O R A R Y   I M P L E M E N T A T I O N
-#
-# THIS IS USED FOR TEST PURPOSES ONLY. DO NOT USE IT OUTSIDE SARDANA TESTS
-#-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-
-
-
-class PropertyInfo():
-
-    def __init__(self, name, type, format, default_value=None):
-        self._name = name
-        self._type = type
-        self._format = format
-        self._default_value = default_value
-
-    def get_name(self):
-        return self._name
-
-    def get_type(self):
-        return self._type
-
-    def get_format(self):
-        return self._format
-
-    def get_default_value(self):
-        return self._default_value
-
-
-#-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-
-# T E M P O R A R Y   I M P L E M E N T A T I O N
-#
-# THIS IS USED FOR TEST PURPOSES ONLY. DO NOT USE IT OUTSIDE SARDANA TESTS
-#-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-
-class ControllerClassInfo(object):
-
-    def __init__(self, name, type, library):
-        self._name = name
-        self._type = type
-        self._library = library
-
-    def get_max_elements(self):
-        return 20
-
-    def get_name(self):
-        return self._name
-
-    def get_model(self):
-        # fake data ###############
-        return "Model of " + self._name
-
-    def get_icon(self):
-        # fake data ###############
-        import taurus.qt.qtgui.resource
-
-        return taurus.qt.qtgui.resource.getIcon(":/designer/extra_motor.png")
-
-    def get_organization(self):
-        # fake data ###############
-        return "Organization of " + self._name
-
-    def get_description(self):
-        #fake data############
-        descr = "This is description of "
-        for i in range(20):
-            descr = descr + " and " + self._name
-        ####################
-        return descr
-
-    def get_family(self):
-        # fake data ###############
-        return "Family of " + self._name
-
-    def get_properties(self):
-        properties = []
-        # fake data ######################
-        properties.append(PropertyInfo(
-            "my parameter", "string", "0D", "deviceName"))
-        properties.append(PropertyInfo("asdsadasd", "integer", "0D", 5))
-        properties.append(PropertyInfo("boollll0", "boolean", "0D", False))
-        properties.append(PropertyInfo("boollll0", "boolean", "0D", True))
-        properties.append(PropertyInfo("boollll0", "boolean", "0D", False))
-        properties.append(PropertyInfo("number1", "float", "0D", 3.5))
-        properties.append(PropertyInfo("string2", "string", "0D", "hehe"))
-        properties.append(PropertyInfo(
-            "tableIntegerD1", "integer", "1D", [1, 2, 3]))
-        properties.append(PropertyInfo(
-            "tablestringD1", "string", "1D", ["aaaa", "bbb", "ccc"]))
-        properties.append(PropertyInfo(
-            "tablefloatD1", "float", "1D", [1.0, 2.5, 3.6]))
-        properties.append(PropertyInfo("tablebooleanD1",
-                                       "boolean", "1D", [True, False, True, False]))
-        properties.append(PropertyInfo(
-            "tablebleintegerD1", "integer", "1D", [1, 2, 3]))
-        properties.append(PropertyInfo("tablebooleanD2", "boolean", "2D", [
-                          [True, False, True], [True, True, True], [False, False, False]]))
-        properties.append(PropertyInfo("tableinteger2", "integer", "2D", [
-                          [1, 2, 3], [11, 22, 33], [-10, -20, -30]]))
-        properties.append(PropertyInfo("tablefloatD2", "float", "2D", [
-                          [0.5, 0.6, 0.8], [0.4, 0.0, 0.333333], [-0.1111, 1, 123123.6]]))
-        properties.append(PropertyInfo("tablestringD2", "string", "2D", [
-                          ["aaaa", "bbb", "ccc"], ["aaaa2", "bbb2", "ccc2"], ["aaaa3", "bbb3", "ccc3"]]))
-
-        return properties
-
-    def get_controller_type(self):
-        return self._type
-
-
-#-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-
-# T E M P O R A R Y   I M P L E M E N T A T I O N
-#
-# THIS IS USED FOR TEST PURPOSES ONLY. DO NOT USE IT OUTSIDE SARDANA TESTS
-#-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-
-class ControllerInfo(object):
-
-    def __init__(self, name, ctrl_class_info):
-        self._name = name
-        self._ctrl_class_info = ctrl_class_info
-
-    def get_controller_class_info(self):
-        return self._ctrl_class_info
-
-    def get_controller_type(self):
-        return self._ctrl_class_info.get_controller_type()
-
-    def get_name(self):
-        return self._name
-
-    def get_max_elements(self):
-        return self._ctrl_class_info.get_max_elements()
-
-    def is_axis_free(self, axis):
-        # fake data
-        if axis == 3:
-            return False
-        else:
-            return True
-
-    def is_name_free(self, name):
-        # fake data
-        if name == "asd":
-            return False
-        else:
-            return True
-
-    def get_icon(self):
-        return self._ctrl_class_info.get_icon()
-
-#-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-
-# T E M P O R A R Y   I M P L E M E N T A T I O N
-#
-# THIS IS USED FOR TEST PURPOSES ONLY. DO NOT USE IT OUTSIDE SARDANA TESTS
-#-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-
-
-
-class Pool(object):
-
-    def __init__(self, sardana, name, poolpath, version, alias=None, device_name=None):
-        self._sardana = sardana
-        self._name = name
-        self._poolpath = poolpath
-        self._version = version
-        self._alias = alias
-        self._device_name = device_name
-
-    def starter_run(self, host, level=1):
-        return True
-
-    def get_name(self):
-        return self._name
-
-    def local_run(self):
-        return True
-
-    def get_element_types(self):
-        return sorted(PoolElementType.keys())
-
-    def get_controller_class_infos(self):
-        #fake data ########################
-        data = []
-        for i in range(5):
-            data.append(ControllerClassInfo("motorController" +
-                                            str(i), PoolElementType.Motor, None))
-        for i in range(5):
-            data.append(ControllerClassInfo(
-                "counterTimerController" + str(i), PoolElementType.CounterTimer, None))
-
-        return data
-
-    def get_controller_infos(self):
-        ctrl_classes = self.get_controller_class_infos()
-        data = []
-        for i in range(2):
-            data.append(ControllerInfo(
-                "My_motor_ctrl_" + str(i), ctrl_classes[i]))
-        for i in range(2):
-            data.append(ControllerInfo(
-                "My_ct_ctrl_" + str(i), ctrl_classes[i + 5]))
-        return data
-
-    def create_controller(self, controller_class_info, name, properties):
-        pass
-
-    def create_element(self, controller_name, name, axis):
-        pass
-
-
-#-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-
-# T E M P O R A R Y   I M P L E M E N T A T I O N
-#
-# THIS IS USED FOR TEST PURPOSES ONLY. DO NOT USE IT OUTSIDE SARDANA TESTS
-#-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-
-class MacroServer(object):
-
-    def __init__(self, sardana, name, macropath, pool_names, version, alias=None, device_name=None):
-        self._sardana = sardana
-        self._name = name
-        self._macropath = macropath
-        self._pool_names = pool_names
-        self._version = version
-        self._alias = alias
-        self._device_name = device_name
-        self._doors = []
-
-    def create_door(self, alias, device_name):
-        try:
-            return self._create_door(alias, device_name)
-        except:
-            db = self.get_database()
-            db.delete_device(device_name)
-            raise
-
-    def _create_door(self, alias, device_name):
-        db = self.get_database()
-        info = PyTango.DbDevInfo()
-        info.name = device_name
-        info._class = "Door"
-        info.server = "MacroServer/" + self._name
-        db.add_device(info)
-        if alias:
-            db.put_device_alias(device_name, alias)
-        door = Door(alias=alias, device_name=device_name)
-        self._doors.append(door)
-        return door
-
-    def remove_door(self, device_name):
-        pass
-
-    def starter_run(self, host, level=1):
-        return True
-
-    def local_run(self):
-        return True
-
-    def get_database(self):
-        return self._sardana.get_database()
-
-#-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-
-# T E M P O R A R Y   I M P L E M E N T A T I O N
-#
-# THIS IS USED FOR TEST PURPOSES ONLY. DO NOT USE IT OUTSIDE SARDANA TESTS
-#-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-
-
-
-class Door(object):
-
-    def __init__(self, alias=None, device_name=None):
-        self._name = alias
-        self._device_name = device_name
-
-
-#-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-
-# T E M P O R A R Y   I M P L E M E N T A T I O N
-#
-# THIS IS USED FOR TEST PURPOSES ONLY. DO NOT USE IT OUTSIDE SARDANA TESTS
-#-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-
-class Sardana(object):
-
-    def __init__(self, sardana_db, name, device_name=None):
-        self._sardana_db = sardana_db
-        self._name = name
-        self._device_name = device_name
-        self._pools = []
-        self._macroservers = []
-        self._init()
-
-    def _init(self):
-        if not self._device_name:
-            return
-        self._pools = []
-        self._macroservers = []
-        dev_name = self._device_name
-        db = self.get_database()
-        cache = db.cache()
-        dev_info = cache.devices()[dev_name]
-        dev_class_name = dev_info.klass().name()
-        if dev_class_name == "Pool":
-            pass
-        elif dev_class_name == "MacroServer":
-            ms_dev_name = dev_name
-            ms_prop_list = map(
-                str.lower, db.get_device_property_list(ms_dev_name, "*"))
-            ms_props = db.get_device_property(ms_dev_name, ms_prop_list)
-            ms_name = dev_info.server().serverInstance()
-            ms_alias = dev_info.alias()
-            ms = MacroServer(self, ms_name, ms_props.get("macropath"), ms_props.get("poolnames"),
-                             ms_props.get("version"), ms_alias, ms_dev_name)
-            self._macroservers.append(ms)
-            for pool_dev_name in ms_props.get("poolnames", ()):
-                pool_prop_list = map(
-                    str.lower, db.get_device_property_list(pool_dev_name, "*"))
-                pool_props = db.get_device_property(
-                    pool_dev_name, pool_prop_list)
-                pool_dev_info = cache.devices()[pool_dev_name]
-                pool_name = pool_dev_info.server().serverInstance()
-                pool_alias = pool_dev_info.alias()
-                pool = Pool(self, pool_name, pool_props.get(
-                    "poolpath"), pool_props.get("version"), pool_alias, pool_dev_name)
-                self._pools.append(pool)
-
-    def get_name(self):
-        return self._name
-
-    def set_device_name(self, device_name):
-        self._device_name = device_name
-        self._init()
-
-    def get_device_name(self):
-        return self._device_name
-
-    def get_pools(self):
-        return self._pools
-
-    def get_macro_servers(self):
-        return self._macro_servers
-
-    def create_pool(self, name, poolpath, version, alias=None, device_name=None):
-        try:
-            return self._create_pool(name, poolpath, version, alias=alias, device_name=device_name)
-        except:
-            db = self.get_database()
-            db.delete_device(device_name)
-            raise
-
-    def _create_pool(self, name, poolpath, version, alias=None, device_name=None):
-        db = self.get_database()
-        info = PyTango.DbDevInfo()
-        info.name = device_name
-        info._class = "Pool"
-        info.server = "Pool/" + name
-        db.add_device(info)
-        if alias:
-            db.put_device_alias(device_name, alias)
-
-        db.put_device_property(
-            device_name, {"PoolPath": poolpath, "Version": version})
-        pool = Pool(self, name, poolpath, version,
-                    alias=alias, device_name=device_name)
-        self._pools.append(pool)
-        db.cache().refresh()
-        return pool
-
-    def create_macroserver(self, name, macropath, pool_names, version, alias=None, device_name=None):
-        try:
-            return self._create_macroserver(name, macropath, pool_names, version, alias=alias, device_name=device_name)
-        except:
-            db = self.get_database()
-            db.delete_device(device_name)
-            raise
-
-    def _create_macroserver(self, name, macropath, pool_names, version, alias=None, device_name=None):
-        db = self.get_database()
-        info = PyTango.DbDevInfo()
-        info.name = device_name
-        info._class = "MacroServer"
-        info.server = "MacroServer/" + name
-        db.add_device(info)
-        if alias:
-            db.put_device_alias(device_name, alias)
-
-        db.put_device_property(device_name, {
-                               "MacroPath": macropath, "Version": version, "PoolNames": pool_names})
-        ms = MacroServer(self, name, macropath, pool_names,
-                         version, alias=alias, device_name=device_name)
-        self._macroservers.append(ms)
-        db.cache().refresh()
-        return ms
-
-    def remove_pool(self):
-        pass
-
-    def remove_macroserver(self):
-        pass
-
-    def get_database(self):
-        return self._sardana_db.get_database()
-
-#-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-
-# T E M P O R A R Y   I M P L E M E N T A T I O N
-#
-# THIS IS USED FOR TEST PURPOSES ONLY. DO NOT USE IT OUTSIDE SARDANA TESTS
-#-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-
-
-
-class DatabaseSardana(object):
-    """A class containning all sardanas for a single database"""
-
-    def __init__(self, db):
-        assert(db is not None)
-        self._db = db
-        self.refresh()
-
-    def refresh(self):
-        self._sardanas = sardanas = {}
-        services = self._db.get_service_list("Sardana/.*")
-        for service, dev in services.items():
-            service_type, service_instance = service.split("/", 1)
-            try:
-                sardanas[service_instance] = Sardana(
-                    self, service_instance, dev)
-            except:
-                pass
-
-    def create_sardana(self, name, device_name):
-        if self._sardanas.has_key(name):
-            raise Exception("Sardana '%s' already exists" % name)
-        self._db.register_service("Sardana", name, device_name)
-        sardana = Sardana(self, name)
-        self._sardanas[name] = sardana
-        return sardana
-
-    def remove_sardana(self, name):
-        try:
-            self._sardanas.pop(name)
-        except KeyError:
-            raise Exception("Sardana '%s' does NOT exist" % name)
-        self._db.unregister_service("Sardana", name)
-
-    def get_sardanas(self):
-        return self._sardanas
-
-    def get_sardana(self, name):
-        return self._sardanas[name]
-
-    def get_database(self):
-        return self._db
-
-#-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-
-# T E M P O R A R Y   I M P L E M E N T A T I O N
-#
-# THIS IS USED FOR TEST PURPOSES ONLY. DO NOT USE IT OUTSIDE SARDANA TESTS
-#-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-
-
-
-class SardanaManager(Singleton, Logger):
-
-    def __init__(self):
-        """ Initialization. Nothing to be done here for now."""
-        pass
-
-    def init(self, *args, **kwargs):
-        """Singleton instance initialization.
-           **For internal usage only**"""
-        name = self.__class__.__name__
-        self.call__init__(Logger, name)
-        self._db_sardanas = {}
-
-    def _get_db_sardana(self, db=None):
-        if db is None:
-            import taurus
-            db = taurus.Database()
-        db_sardana = self._db_sardanas.get(db)
-        if db_sardana is None:
-            db_sardana = DatabaseSardana(db)
-        return db_sardana
-
-    def create_sardana(self, name, device_name, db=None):
-        return self._get_db_sardana(db).create_sardana(name, device_name)
-
-    def remove_sardana(self, name, db=None):
-        self._get_db_sardana(db).remove_sardana(name)
-
-    def get_sardanas(self, db=None):
-        return self._get_db_sardana(db).get_sardanas()
-
-    def get_sardana(self, name, db=None):
-        return self._get_db_sardana(db).get_sardana(name)
-
-    def get_hosts(self):
-        return ["localhost"] + ["controls%02d" % i for i in range(5)]
-
-    def get_level_range(self):
-        return 1, 200
-
-    def has_localhost_starter(self):
-        import socket
-        return socket.gethostname() in self.get_hosts()
-
-    @classmethod
-    def get_default_pool_path(cls):
-        pathList = []
-        pathList.append("/homelocal/sicilia/lib/poolcontrollers")
-        pathList.append(
-            "/homelocal/sicilia/lib/python/site-packages/poolcontrollers")
-        return pathList
-
-    @classmethod
-    def get_default_ms_path(cls):
-        pathList = []
-        pathList.append(
-            "/homelocal/sicilia/lib/python/site-packages/macroserver/macros")
-        return pathList
